+++ src/frc/team1983/utilities/sensors/Encoder.java
@@ -3,6 +3,6 @@ package frc.team1983.utilities.sensors;
 public interface Encoder
 {
     void configure();
-    int getPosition();
+    double getPosition();
     double getVelocity();
 }
